backend/requirements.txt
# FastAPI Backend for MarketGPS
# Handles Stripe webhooks and billing endpoints

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
stripe>=7.0.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
pyarrow>=14.0.0
python-multipart>=0.0.9
